pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "numpy>=2.3.0",
    "pandas>=2.3.0",
    "pillow>=11.2.1",
    "plotly>=6.1.2",
    "requests>=2.32.4",
    "scikit-learn>=1.7.0",
    "streamlit>=1.45.1",
    "trafilatura>=2.0.0",
]
